ui: Calculate gaps delta based on 1000 samples per graph
If we have more than 2 missing samples we add a gap into the graph
