+++ ui/src/components/graphs/colors.tsx
